# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rhytmiq API",
    "description": "FIG-compliant API for managing rhythmic gymnastics meets",
    "version": "0.1.0"
  },
  "paths": {
...
    "/routines/": {
      "post": {
        "tags": [
          "Routines"
        ],
        "summary": "Create Routine",
        "operationId": "create_routine_routines__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Routines"
        ],
        "summary": "List Routines",
        "operationId": "list_routines_routines__get",
        "parameters": [
          {
            "name": "entry_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entry_id",
              "title": "Entry Id"
            },
            "description": "Filter by entry_id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutineRead"
                  },
                  "title": "Response List Routines Routines  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}": {
      "get": {
        "tags": [
          "Routines"
        ],
        "summary": "Get Routine",
        "operationId": "get_routine_routines__routine_id__get",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Routines"
        ],
        "summary": "Update Routine",
        "operationId": "update_routine_routines__routine_id__patch",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Routines"
        ],
        "summary": "Delete Routine",
        "operationId": "delete_routine_routines__routine_id__delete",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/routines/{routine_id}/score": {
      "get": {
        "tags": [
          "Routines"
        ],
        "summary": "Get Routine Score",
        "operationId": "get_routine_score_routines__routine_id__score_get",
        "parameters": [
          {
            "name": "routine_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Routine Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutineScoreRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Apparatus": {
        "type": "string",
        "enum": [
          "rope",
          "hoop",
          "ball",
          "clubs",
          "ribbon",
          "freehand"
        ],
        "title": "Apparatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoutineCreate": {
        "properties": {
          "entry_id": {
            "type": "integer",
            "title": "Entry Id",
            "description": "The ID of the meet entry this routine belongs to."
          },
          "apparatus": {
            "$ref": "#/components/schemas/Apparatus",
            "description": "The apparatus for this routine."
          },
          "order_of_performance": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Of Performance",
            "description": "The order of performance for this routine."
          },
          "penalty": {
            "anyOf": [
              {
                "type": "number",
                "multipleOf": 0.05,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Penalty",
            "description": "Deduction applied to this routine's final score.",
            "default": "0"
          }
        },
        "type": "object",
        "required": [
          "entry_id",
          "apparatus"
        ],
        "title": "RoutineCreate"
      },
...
      "RoutineRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "The unique identifier for this routine."
          },
          "entry_id": {
            "type": "integer",
            "title": "Entry Id",
            "description": "The ID of the meet entry this routine belongs to."
          },
          "apparatus": {
            "$ref": "#/components/schemas/Apparatus",
            "description": "The apparatus for this routine."
          },
          "order_of_performance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Of Performance",
            "description": "The order of performance for this routine."
          },
          "penalty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Penalty",
            "description": "Deduction applied to this routine's final score."
          }
        },
        "type": "object",
        "required": [
          "id",
          "entry_id",
          "apparatus",
          "penalty"
        ],
        "title": "RoutineRead"
      },
      "RoutineScoreRead": {
        "properties": {
          "routine_id": {
            "type": "integer",
            "title": "Routine Id",
            "description": "The ID of the routine this score belongs to."
          },
          "d_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "D Score",
            "description": "Difficulty score: difficulty_body + difficulty_apparatus subgroups summed."
          },
          "a_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "A Score",
            "description": "Artistry score (trimmed mean of artistry marks)."
          },
          "e_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "E Score",
            "description": "Execution score (trimmed mean of execution marks)."
          },
          "final_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Final Score",
            "description": "Levels 1-3 only: the single pre-aggregated mark out of 13. 0 at every other level, where the score is built from the D/A/E panels instead."
          },
          "penalty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Penalty",
            "description": "Deduction applied to this routine's final score."
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total",
            "description": "d_score + a_score + e_score - penalty."
          }
        },
        "type": "object",
        "required": [
          "routine_id",
          "d_score",
          "a_score",
          "e_score",
          "final_score",
          "penalty",
          "total"
        ],
        "title": "RoutineScoreRead",
        "description": "The live-computed D/A/E/total score for a routine (see app/scoring.py)."
      },
      "RoutineUpdate": {
        "properties": {
          "order_of_performance": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Of Performance",
            "description": "The order of performance for this routine."
          },
          "penalty": {
            "anyOf": [
              {
                "type": "number",
                "multipleOf": 0.05,
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Penalty",
            "description": "Deduction applied to this routine's final score."
          }
        },
        "type": "object",
        "title": "RoutineUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}